Revert "Revert "Refactored prim utils into _prims_utils folder (#81088)""
This reverts commit e43a02c3140d0cc00e43b3a1e831d5fb9ee68ceb.

[ghstack-poisoned]
