+++ .github/ci_commit_pins/torchdynamo.txt
@@ -1,1 +1,1 @@
-dbb005e7429300e6605abdf6533d0d6dac8dabe3
+30ab3e6cb3c39678c366f102a49cfb59f8507b15
