+++ test/test_fake_tensor.py
@@ -315,13 +315,13 @@ class FakeTensorConverterTest(TestCase):
 
     def test_no_ref_cycle(self):
         x = torch.rand([4])
-        mode = torch._prims.utils.get_prim_fake_mode()
+        mode = torch._prims.get_prim_fake_mode()
         y = mode.from_tensor(x)
-        assert mode is torch._prims.utils.get_prim_fake_mode()
+        assert mode is torch._prims.get_prim_fake_mode()
         self.assertEqual(len(mode.fake_tensor_converter.tensor_memo), 1)
         del mode
         del y
-        new_mode = torch._prims.utils.get_prim_fake_mode()
+        new_mode = torch._prims.get_prim_fake_mode()
         self.assertEqual(len(new_mode.fake_tensor_converter.tensor_memo), 0)
