+++ torch/_decomp/decompositions.py
@@ -6,8 +6,8 @@ from typing import Tuple, Optional, List, Callable
 import torch.nn.functional as F
 import functools
 from torch.utils._pytree import tree_map, tree_flatten
-import torch._prims.utils as utils
-from torch._prims.wrappers import out_wrapper
+import torch._prims_common as utils
+from torch._prims_common.wrappers import out_wrapper
 
 # None of these functions are publicly accessible; get at them
 # from torch._decomps
